moved dart files into folders for models, pages and widgets
